Feature/metodo pagamento crud

+++ src/main/java/com/controle_de_gastos/notas_api/service/MetodoPagamentoService.java
@@ -1,9 +1,12 @@
 package com.controle_de_gastos.notas_api.service;
 
+import com.controle_de_gastos.notas_api.dto.projecao.NotaSimplesProjecaoDTO;
 import com.controle_de_gastos.notas_api.dto.requisicao.MetodoPagamentoRequisicaoDTO;
+import com.controle_de_gastos.notas_api.dto.resposta.MetodoPagamentoComNotaRespostaDTO;
 import com.controle_de_gastos.notas_api.repository.MetodoPagamentoRepository;
 import com.controle_de_gastos.notas_api.dto.resposta.MetodoPagamentoRespostaDTO;
 import com.controle_de_gastos.notas_api.model.MetodoPagamento;
+import com.controle_de_gastos.notas_api.repository.NotaMetodoPagametoJuncaoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -14,8 +17,9 @@ import java.util.Optional;
 public class MetodoPagamentoService {
 
    private final MetodoPagamentoRepository metodoPagamentoRepository;
+    private final NotaMetodoPagametoJuncaoRepository notaMetodoPagametoJuncaoRepository;
 
-   public MetodoPagamentoRespostaDTO toRespostaDTO(MetodoPagamento metodoPagamento) {
+    public MetodoPagamentoRespostaDTO toRespostaDTO(MetodoPagamento metodoPagamento) {
        return new MetodoPagamentoRespostaDTO(
                metodoPagamento.getId(),
                metodoPagamento.getNome());
@@ -33,13 +37,67 @@ public class MetodoPagamentoService {
                .map(this::toRespostaDTO);
    }
 
+   public Optional<MetodoPagamentoComNotaRespostaDTO> listarNotasPorMetodoId(Integer id){
+       Optional<MetodoPagamento> metodo = metodoPagamentoRepository.findById(id);
+       if(metodo.isEmpty())return Optional.empty();
+
+       List<NotaSimplesProjecaoDTO> notas = notaMetodoPagametoJuncaoRepository.findByMetodoPagamentoId(id)
+               .stream()
+               .map(juncao -> new NotaSimplesProjecaoDTO(
+                       juncao.getNota().getId(),
+                       juncao.getNota().getData(),
+                       juncao.getNota().getTotal()
+               ))
+               .toList();
+
+        return Optional.of(new MetodoPagamentoComNotaRespostaDTO(
+                metodo.get().getId(),
+                metodo.get().getNome(),
+                notas
+        ));
+   }
+
+   public List<MetodoPagamentoComNotaRespostaDTO> listarNotasPorMetodoPagamentoTodos(){
+        return metodoPagamentoRepository.findAll()
+                .stream()
+                .map(metodo -> new MetodoPagamentoComNotaRespostaDTO(
+                        metodo.getId(),
+                        metodo.getNome(),
+                        metodo.getNotaMetodoPagamentoJuncaos()
+                                .stream()
+                                .map(juncao -> new NotaSimplesProjecaoDTO(
+                                        juncao.getNota().getId(),
+                                        juncao.getNota().getData(),
+                                        juncao.getNota().getTotal()
+                                ))
+                                .toList()
+                ))
+                .toList();
+
+   }
+
    public MetodoPagamentoRespostaDTO criar(MetodoPagamentoRequisicaoDTO metodoDTO){
-       MetodoPagamento metodoPagamento = new MetodoPagamento();
-       metodoPagamento.setNome(metodoDTO.nome());
+       MetodoPagamento metodoPagamento = MetodoPagamento.builder()
+               .nome(metodoDTO.nome())
+               .build();
        return toRespostaDTO(metodoPagamentoRepository.save(metodoPagamento));
    }
 
-   public void deletarPorId(Integer id){
+   public Optional<MetodoPagamentoRespostaDTO> atualizarTudo(Integer id, MetodoPagamentoRequisicaoDTO metodoDTO){
+       return metodoPagamentoRepository.findById(id)
+               .map(metodo -> {
+                   metodo.setNome(metodoDTO.nome());
+                   return toRespostaDTO(metodoPagamentoRepository.save(metodo));
+               });
+   }
+
+   public boolean deletarPorId(Integer id){
+       MetodoPagamento metodoPagamento = metodoPagamentoRepository.findById(id).orElse(null);
+       if (metodoPagamento == null) return false;
+
+       if(!metodoPagamento.getNotaMetodoPagamentoJuncaos().isEmpty()) throw new IllegalStateException("Não é possível deletar, existem dependências");
+
        metodoPagamentoRepository.deleteById(id);
+       return true;
    }
 }
